+++ src/main/java/ttjj/db/StockAdrCount.java
@@ -94,7 +94,7 @@ public class StockAdrCount {
     /**
      * f21 流通市值
      */
-    private Long f21;
+    private BigDecimal f21;
     /**
      * f62 主力净流入
      */
@@ -367,11 +367,11 @@ public class StockAdrCount {
         this.f20 = f20;
     }
 
-    public Long getF21() {
+    public BigDecimal getF21() {
         return f21;
     }
 
-    public void setF21(Long f21) {
+    public void setF21(BigDecimal f21) {
         this.f21 = f21;
     }
